requirements.txt
torch>=1.9.0
fair-esm>=2.0.0
biopython>=1.79
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
tqdm>=4.64.0
tensorboard>=2.8.0
numpy>=1.21.0
pandas>=1.3.0
